+++ otterdog/models/environment_variable.py
@@ -1,0 +1,71 @@
+#  *******************************************************************************
+#  Copyright (c) 2026 Eclipse Foundation and others.
+#  This program and the accompanying materials are made available
+#  under the terms of the Eclipse Public License 2.0
+#  which is available at http://www.eclipse.org/legal/epl-v20.html
+#  SPDX-License-Identifier: EPL-2.0
+#  *******************************************************************************
+
+from __future__ import annotations
+
+import dataclasses
+from typing import TYPE_CHECKING
+
+from otterdog.models import LivePatch, LivePatchType
+from otterdog.models.variable import Variable
+from otterdog.utils import expect_type, unwrap
+
+if TYPE_CHECKING:
+    from otterdog.jsonnet import JsonnetConfig
+    from otterdog.providers.github import GitHubProvider
+
+
+@dataclasses.dataclass
+class EnvironmentVariable(Variable):
+    """
+    Represents a Variable defined on environment level.
+    """
+
+    @property
+    def model_object_name(self) -> str:
+        return "env_variable"
+
+    def get_jsonnet_template_function(self, jsonnet_config: JsonnetConfig, extend: bool) -> str | None:
+        return f"orgs.{jsonnet_config.create_environment_variable}"
+
+    @classmethod
+    async def apply_live_patch(
+        cls,
+        patch: LivePatch[EnvironmentVariable],
+        org_id: str,
+        provider: GitHubProvider,
+    ) -> None:
+        from .environment import Environment
+
+        environment = expect_type(patch.parent_object, Environment)
+
+        match patch.patch_type:
+            case LivePatchType.ADD:
+                await provider.add_environment_variable(
+                    org_id,
+                    environment.repo_name,
+                    environment.name,
+                    await unwrap(patch.expected_object).to_provider_data(org_id, provider),
+                )
+
+            case LivePatchType.REMOVE:
+                await provider.delete_environment_variable(
+                    org_id,
+                    environment.repo_name,
+                    environment.name,
+                    unwrap(patch.current_object).name,
+                )
+
+            case LivePatchType.CHANGE:
+                await provider.update_environment_variable(
+                    org_id,
+                    environment.repo_name,
+                    environment.name,
+                    unwrap(patch.current_object).name,
+                    await unwrap(patch.expected_object).to_provider_data(org_id, provider),
+                )