+++ otterdog/providers/github/__init__.py
@@ -479,6 +479,36 @@ class GitHubProvider:
     async def delete_repo_variable(self, org_id: str, repo_name: str, variable_name: str) -> None:
         await self.rest_api.repo.delete_variable(org_id, repo_name, variable_name)
 
+    async def get_environment_secrets(self, org_id: str, repo_name: str, env_name: str) -> list[dict[str, Any]]:
+        return await self.rest_api.repo.get_environment_secrets(org_id, repo_name, env_name)
+
+    async def update_environment_secret(
+        self, org_id: str, repo_name: str, env_name: str, secret_name: str, secret: dict[str, Any]
+    ) -> None:
+        if len(secret) > 0:
+            await self.rest_api.repo.update_environment_secret(org_id, repo_name, env_name, secret_name, secret)
+
+    async def add_environment_secret(self, org_id: str, repo_name: str, env_name: str, data: dict[str, str]) -> None:
+        await self.rest_api.repo.add_environment_secret(org_id, repo_name, env_name, data)
+
+    async def delete_environment_secret(self, org_id: str, repo_name: str, env_name: str, secret_name: str) -> None:
+        await self.rest_api.repo.delete_environment_secret(org_id, repo_name, env_name, secret_name)
+
+    async def get_environment_variables(self, org_id: str, repo_name: str, env_name: str) -> list[dict[str, Any]]:
+        return await self.rest_api.repo.get_environment_variables(org_id, repo_name, env_name)
+
+    async def update_environment_variable(
+        self, org_id: str, repo_name: str, env_name: str, variable_name: str, variable: dict[str, Any]
+    ) -> None:
+        if len(variable) > 0:
+            await self.rest_api.repo.update_environment_variable(org_id, repo_name, env_name, variable_name, variable)
+
+    async def add_environment_variable(self, org_id: str, repo_name: str, env_name: str, data: dict[str, str]) -> None:
+        await self.rest_api.repo.add_environment_variable(org_id, repo_name, env_name, data)
+
+    async def delete_environment_variable(self, org_id: str, repo_name: str, env_name: str, variable_name: str) -> None:
+        await self.rest_api.repo.delete_environment_variable(org_id, repo_name, env_name, variable_name)
+
     async def get_team_permissions(self, org_id: str) -> list[dict[str, Any]]:
         return await self.graphql_client.get_team_permissions(org_id)